Adds thank you page after form submission
Implements a thank you page to confirm successful form submissions.

This commit introduces a new "thank-you-page" route and component, displayed after a user submits the questions page form. It provides confirmation that the submission was successful and informs the user of the next steps. Also, it adds the guest type to the questions page so that the correct redirect can be made. Finally, it removes the check to see if the guest has already been invited.